Moved Dashboard User Details into separate component
