+++ src/components/dashboard/my-details/index.js
@@ -1,0 +1,3 @@
+import MyDetails from "./mydetailscontroller"
+
+export default MyDetails
